adjusted imaging stitching to account for ranges
